Set default pagesize for KDB and PTree to the values suggested by db_tuner
Set the default pagesize for KDB to 65536 and for PTree to 4096. These are
the values recommended by Oracle's db_tuner utility.

Correct usage info for -pagesize and -ptree_pagesize in settings.ml and sks.pod to reflect that the value of these is the number of 512byte blocks to allocate,
not the number of bytes.