make ocextr as bytecode rather than native
ocextr is only ever run at build time, and it is not
performance-sensitive.  On some architectures, ocaml can *only*
produce bytecode; on those architectures, we'll still need ocextr in
order to complete the build.

On architectures where ocaml can produce native code, it still works
as bytecode, so this improvement doesn't break anything.